Added a database
