+++ src/main/resources/application.properties
@@ -1,1 +1,6 @@
-
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.url=jdbc:mysql://localhost:3306/bank_system
+spring.datasource.username=root
+spring.datasource.password=differential1$
+spring.datasource.driver-class-name =com.mysql.jdbc.Driver
+#spring.jpa.show-sql: true
